+++ function.cpp
@@ -28,6 +28,7 @@ void FunctionC::init()
   m_expstrAnalyzed = false;
   m_fieldnames = NULL;
   m_fieldtypes = NULL;
+  m_anaGroupNum = 0;
 
   m_metaDataAnzlyzed = false; // analyze column name to column id.
 }
@@ -106,23 +107,63 @@ bool FunctionC::analyzeExpStr()
   //  return false;
   //}
   m_funcName = trim_copy(upper_copy(m_expStr.substr(0, m_expStr.find("("))));
+  m_funcID = encodeFunction(m_funcName);
   m_expStr = m_funcName+"("+strParams+")";
   //strParams = trim_pair(strParams, "()");
   vector<string> vParams = split(strParams,',',"''()",'\\',{'(',')'},false,true);
-  for (int i=0; i<vParams.size(); i++){
-    trace(DEBUG, "Processing parameter(%d) '%s'!\n", i, vParams[i].c_str());
-    string sParam = trim_copy(vParams[i]);
-    if (sParam.empty()){
-      trace(ERROR, "Empty parameter string!\n");
+  if (isAnalytic()){ // analyze analytic function parameters: (group1[;group2]...[,sort1 [asc|desc][;sort2 [asc|desc]]...]). sParam of analytic function can be empty
+    if (vParams.size()!=1&&vParams.size()!=2){
+      trace(ERROR, "There should be two part parameters for an analytic function '%s', a comma should be provided even no value applied, for example rank(,)!\n",m_funcName.c_str());
       m_expstrAnalyzed = false;
       return false;
     }
-    ExpressionC eParam = ExpressionC(sParam);
-    //trace(DEBUG2,"'%s' merged const to '%s'.\n",sParam.c_str(),eParam.getEntireExpstr().c_str());
-    //eParam.analyzeColumns(m_fieldnames, m_fieldtypes, rawDatatype);
-    m_params.push_back(eParam);
+    vector<string> vAnaPara = split(trim_copy(vParams[0]),';',"''()",'\\',{'(',')'},false,true);
+    if (vAnaPara.size()==0){ // need a param for group, if not provided, give it a ""
+      m_anaGroupNum = 1;
+      ExpressionC eParam = ExpressionC("");
+      m_params.push_back(eParam);
+      eParam = ExpressionC("1"); // group always get asc
+      m_params.push_back(eParam);
+    }else{
+      m_anaGroupNum = vAnaPara.size();
+      for (int i=0;i<vAnaPara.size();i++){
+        ExpressionC eParam = ExpressionC(trim_copy(vAnaPara[i]));
+        m_params.push_back(eParam);
+        eParam = ExpressionC("1"); // group always get asc
+        m_params.push_back(eParam);
+      }
+    }
+    if (vParams.size()==2){
+      vAnaPara = split(trim_copy(vParams[1]),';',"''()",'\\',{'(',')'},false,true);
+      for (int i=0;i<vAnaPara.size();i++){
+        vector<string> vSortPara = split(trim_copy(vAnaPara[i]),' ',"''()",'\\',{'(',')'},true,true);
+        ExpressionC eParam = ExpressionC(trim_copy(vSortPara[0]));
+        m_params.push_back(eParam);
+        if (vSortPara.size()>1){ // sort direction; 1:asc;-1:desc
+          eParam = ExpressionC(upper_copy(trim_copy(vSortPara[1])).compare("DESC")==0?"-1":"1");
+          m_params.push_back(eParam);
+        }else{
+          eParam = ExpressionC("1"); // default is asc
+          m_params.push_back(eParam);
+        }
+      }
+    }
+    trace(DEBUG, "FunctionC: The analytic function '%s' group size is %d, param size %d \n", m_expStr.c_str(), m_anaGroupNum, m_params.size());
+  }else{
+    for (int i=0; i<vParams.size(); i++){
+      trace(DEBUG, "Processing parameter(%d) '%s'!\n", i, vParams[i].c_str());
+      string sParam = trim_copy(vParams[i]);
+      if (sParam.empty()){
+        trace(ERROR, "Empty parameter string!\n");
+        m_expstrAnalyzed = false;
+        return false;
+      }
+      ExpressionC eParam = ExpressionC(sParam);
+      //trace(DEBUG2,"'%s' merged const to '%s'.\n",sParam.c_str(),eParam.getEntireExpstr().c_str());
+      //eParam.analyzeColumns(m_fieldnames, m_fieldtypes, rawDatatype);
+      m_params.push_back(eParam);
+    }
   }
-  m_funcID = encodeFunction(m_funcName);
   switch(m_funcID){
     case UPPER:
     case LOWER:
@@ -228,6 +269,7 @@ FunctionC* FunctionC::cloneMe(){
   node->m_funcName = m_funcName;
   node->m_funcID = m_funcID;
   node->m_params = m_params;
+  node->m_anaGroupNum = m_anaGroupNum;
   node->m_fieldnames = m_fieldnames;
   node->m_fieldtypes = m_fieldtypes;
 
@@ -245,6 +287,7 @@ void FunctionC::copyTo(FunctionC* node){
     node->m_funcName = m_funcName;
     node->m_funcID = m_funcID;
     node->m_params = m_params;
+    node->m_anaGroupNum = m_anaGroupNum;
     node->m_fieldnames = m_fieldnames;
     node->m_fieldtypes = m_fieldtypes;
   }
@@ -262,6 +305,8 @@ void FunctionC::clear(){
   m_expstrAnalyzed = false;
   m_fieldnames = NULL;
   m_fieldtypes = NULL;
+  m_anaGroupNum = 0;
+  init();
 }
 
 // remove a node from prediction. Note: the input node is the address of the node contains in current prediction